@@ -27,6 +27,7 @@ from . import _transformers as t
 from . import types
 from ._api_client import BaseApiClient
 from ._common import get_value_by_path as getv
+from ._common import move_value_by_path as movev
 from ._common import set_value_by_path as setv
 from .pagers import AsyncPager, Pager
 
@@ -63,7 +64,7 @@ def _BatchJobDestination_from_mldev(
         to_object,
         ['inlined_embed_content_responses'],
         [
-            _InlinedEmbedContentResponse_from_mldev(item, to_object)
+            item
             for item in getv(
                 from_object,
                 ['inlinedEmbedContentResponses', 'inlinedResponses'],
@@ -291,11 +292,7 @@ def _BatchJob_from_vertex(
     setv(to_object, ['state'], t.t_job_state(getv(from_object, ['state'])))
 
   if getv(from_object, ['error']) is not None:
-    setv(
-        to_object,
-        ['error'],
-        _JobError_from_vertex(getv(from_object, ['error']), to_object),
-    )
+    setv(to_object, ['error'], getv(from_object, ['error']))
 
   if getv(from_object, ['createTime']) is not None:
     setv(to_object, ['create_time'], getv(from_object, ['createTime']))
@@ -334,31 +331,16 @@ def _BatchJob_from_vertex(
   return to_object
 
 
-def _Blob_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-
-  if getv(from_object, ['data']) is not None:
-    setv(to_object, ['data'], getv(from_object, ['data']))
-
-  if getv(from_object, ['mimeType']) is not None:
-    setv(to_object, ['mime_type'], getv(from_object, ['mimeType']))
-
-  return to_object
-
-
 def _Blob_to_mldev(
     from_object: Union[dict[str, Any], object],
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
-  if getv(from_object, ['display_name']) is not None:
-    raise ValueError('display_name parameter is not supported in Gemini API.')
-
   if getv(from_object, ['data']) is not None:
     setv(to_object, ['data'], getv(from_object, ['data']))
+
+  if getv(from_object, ['display_name']) is not None:
+    raise ValueError('display_name parameter is not supported in Gemini API.')
 
   if getv(from_object, ['mime_type']) is not None:
     setv(to_object, ['mimeType'], getv(from_object, ['mime_type']))
@@ -404,11 +386,7 @@ def _Candidate_from_mldev(
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
   if getv(from_object, ['content']) is not None:
-    setv(
-        to_object,
-        ['content'],
-        _Content_from_mldev(getv(from_object, ['content']), to_object),
-    )
+    setv(to_object, ['content'], getv(from_object, ['content']))
 
   if getv(from_object, ['citationMetadata']) is not None:
     setv(
@@ -424,15 +402,6 @@ def _Candidate_from_mldev(
 
   if getv(from_object, ['finishReason']) is not None:
     setv(to_object, ['finish_reason'], getv(from_object, ['finishReason']))
-
-  if getv(from_object, ['urlContextMetadata']) is not None:
-    setv(
-        to_object,
-        ['url_context_metadata'],
-        _UrlContextMetadata_from_mldev(
-            getv(from_object, ['urlContextMetadata']), to_object
-        ),
-    )
 
   if getv(from_object, ['avgLogprobs']) is not None:
     setv(to_object, ['avg_logprobs'], getv(from_object, ['avgLogprobs']))
@@ -451,7 +420,18 @@ def _Candidate_from_mldev(
     setv(to_object, ['logprobs_result'], getv(from_object, ['logprobsResult']))
 
   if getv(from_object, ['safetyRatings']) is not None:
-    setv(to_object, ['safety_ratings'], getv(from_object, ['safetyRatings']))
+    setv(
+        to_object,
+        ['safety_ratings'],
+        [item for item in getv(from_object, ['safetyRatings'])],
+    )
+
+  if getv(from_object, ['urlContextMetadata']) is not None:
+    setv(
+        to_object,
+        ['url_context_metadata'],
+        getv(from_object, ['urlContextMetadata']),
+    )
 
   return to_object
 
@@ -462,39 +442,11 @@ def _CitationMetadata_from_mldev(
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
   if getv(from_object, ['citationSources']) is not None:
-    setv(to_object, ['citations'], getv(from_object, ['citationSources']))
-
-  return to_object
-
-
-def _ContentEmbedding_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['values']) is not None:
-    setv(to_object, ['values'], getv(from_object, ['values']))
-
-  return to_object
-
-
-def _Content_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['parts']) is not None:
     setv(
         to_object,
-        ['parts'],
-        [
-            _Part_from_mldev(item, to_object)
-            for item in getv(from_object, ['parts'])
-        ],
+        ['citations'],
+        [item for item in getv(from_object, ['citationSources'])],
     )
-
-  if getv(from_object, ['role']) is not None:
-    setv(to_object, ['role'], getv(from_object, ['role']))
 
   return to_object
 
@@ -714,11 +666,7 @@ def _DeleteResourceJob_from_mldev(
     setv(to_object, ['done'], getv(from_object, ['done']))
 
   if getv(from_object, ['error']) is not None:
-    setv(
-        to_object,
-        ['error'],
-        _JobError_from_mldev(getv(from_object, ['error']), to_object),
-    )
+    setv(to_object, ['error'], getv(from_object, ['error']))
 
   return to_object
 
@@ -740,29 +688,7 @@ def _DeleteResourceJob_from_vertex(
     setv(to_object, ['done'], getv(from_object, ['done']))
 
   if getv(from_object, ['error']) is not None:
-    setv(
-        to_object,
-        ['error'],
-        _JobError_from_vertex(getv(from_object, ['error']), to_object),
-    )
-
-  return to_object
-
-
-def _DynamicRetrievalConfig_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['mode']) is not None:
-    setv(to_object, ['mode'], getv(from_object, ['mode']))
-
-  if getv(from_object, ['dynamic_threshold']) is not None:
-    setv(
-        to_object,
-        ['dynamicThreshold'],
-        getv(from_object, ['dynamic_threshold']),
-    )
+    setv(to_object, ['error'], getv(from_object, ['error']))
 
   return to_object
 
@@ -777,16 +703,21 @@ def _EmbedContentBatch_to_mldev(
     setv(
         to_object,
         ['requests[]', 'request', 'content'],
-        t.t_contents_for_embed(api_client, getv(from_object, ['contents'])),
+        [
+            item
+            for item in t.t_contents_for_embed(
+                api_client, getv(from_object, ['contents'])
+            )
+        ],
     )
 
   if getv(from_object, ['config']) is not None:
     setv(
         to_object,
-        ['config'],
+        ['_self'],
         _EmbedContentConfig_to_mldev(getv(from_object, ['config']), to_object),
     )
-
+    movev(to_object, {'requests[].*': 'requests[].request.*'})
   return to_object
 
 
@@ -843,21 +774,6 @@ def _EmbeddingsBatchJobSource_to_mldev(
   return to_object
 
 
-def _FileData_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-
-  if getv(from_object, ['fileUri']) is not None:
-    setv(to_object, ['file_uri'], getv(from_object, ['fileUri']))
-
-  if getv(from_object, ['mimeType']) is not None:
-    setv(to_object, ['mime_type'], getv(from_object, ['mimeType']))
-
-  return to_object
-
-
 def _FileData_to_mldev(
     from_object: Union[dict[str, Any], object],
     parent_object: Optional[dict[str, Any]] = None,
@@ -871,95 +787,6 @@ def _FileData_to_mldev(
 
   if getv(from_object, ['mime_type']) is not None:
     setv(to_object, ['mimeType'], getv(from_object, ['mime_type']))
-
-  return to_object
-
-
-def _FunctionCall_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['id']) is not None:
-    setv(to_object, ['id'], getv(from_object, ['id']))
-
-  if getv(from_object, ['args']) is not None:
-    setv(to_object, ['args'], getv(from_object, ['args']))
-
-  if getv(from_object, ['name']) is not None:
-    setv(to_object, ['name'], getv(from_object, ['name']))
-
-  return to_object
-
-
-def _FunctionCall_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['id']) is not None:
-    setv(to_object, ['id'], getv(from_object, ['id']))
-
-  if getv(from_object, ['args']) is not None:
-    setv(to_object, ['args'], getv(from_object, ['args']))
-
-  if getv(from_object, ['name']) is not None:
-    setv(to_object, ['name'], getv(from_object, ['name']))
-
-  return to_object
-
-
-def _FunctionCallingConfig_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['mode']) is not None:
-    setv(to_object, ['mode'], getv(from_object, ['mode']))
-
-  if getv(from_object, ['allowed_function_names']) is not None:
-    setv(
-        to_object,
-        ['allowedFunctionNames'],
-        getv(from_object, ['allowed_function_names']),
-    )
-
-  return to_object
-
-
-def _FunctionDeclaration_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['behavior']) is not None:
-    setv(to_object, ['behavior'], getv(from_object, ['behavior']))
-
-  if getv(from_object, ['description']) is not None:
-    setv(to_object, ['description'], getv(from_object, ['description']))
-
-  if getv(from_object, ['name']) is not None:
-    setv(to_object, ['name'], getv(from_object, ['name']))
-
-  if getv(from_object, ['parameters']) is not None:
-    setv(to_object, ['parameters'], getv(from_object, ['parameters']))
-
-  if getv(from_object, ['parameters_json_schema']) is not None:
-    setv(
-        to_object,
-        ['parametersJsonSchema'],
-        getv(from_object, ['parameters_json_schema']),
-    )
-
-  if getv(from_object, ['response']) is not None:
-    setv(to_object, ['response'], getv(from_object, ['response']))
-
-  if getv(from_object, ['response_json_schema']) is not None:
-    setv(
-        to_object,
-        ['responseJsonSchema'],
-        getv(from_object, ['response_json_schema']),
-    )
 
   return to_object
 
@@ -1036,10 +863,7 @@ def _GenerateContentConfig_to_mldev(
     setv(
         to_object,
         ['responseSchema'],
-        _Schema_to_mldev(
-            t.t_schema(api_client, getv(from_object, ['response_schema'])),
-            to_object,
-        ),
+        t.t_schema(api_client, getv(from_object, ['response_schema'])),
     )
 
   if getv(from_object, ['response_json_schema']) is not None:
@@ -1078,11 +902,7 @@ def _GenerateContentConfig_to_mldev(
     )
 
   if getv(from_object, ['tool_config']) is not None:
-    setv(
-        parent_object,
-        ['toolConfig'],
-        _ToolConfig_to_mldev(getv(from_object, ['tool_config']), to_object),
-    )
+    setv(parent_object, ['toolConfig'], getv(from_object, ['tool_config']))
 
   if getv(from_object, ['labels']) is not None:
     raise ValueError('labels parameter is not supported in Gemini API.')
@@ -1112,9 +932,7 @@ def _GenerateContentConfig_to_mldev(
     setv(
         to_object,
         ['speechConfig'],
-        _SpeechConfig_to_mldev(
-            t.t_speech_config(getv(from_object, ['speech_config'])), to_object
-        ),
+        t.t_speech_config(getv(from_object, ['speech_config'])),
     )
 
   if getv(from_object, ['audio_timestamp']) is not None:
@@ -1123,13 +941,10 @@ def _GenerateContentConfig_to_mldev(
     )
 
   if getv(from_object, ['thinking_config']) is not None:
-    setv(
-        to_object,
-        ['thinkingConfig'],
-        _ThinkingConfig_to_mldev(
-            getv(from_object, ['thinking_config']), to_object
-        ),
-    )
+    setv(to_object, ['thinkingConfig'], getv(from_object, ['thinking_config']))
+
+  if getv(from_object, ['image_config']) is not None:
+    setv(to_object, ['imageConfig'], getv(from_object, ['image_config']))
 
   return to_object
 
@@ -1201,19 +1016,16 @@ def _GetBatchJobParameters_to_vertex(
   return to_object
 
 
-def _GoogleSearchRetrieval_to_mldev(
+def _GoogleMaps_to_mldev(
     from_object: Union[dict[str, Any], object],
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
-  if getv(from_object, ['dynamic_retrieval_config']) is not None:
-    setv(
-        to_object,
-        ['dynamicRetrievalConfig'],
-        _DynamicRetrievalConfig_to_mldev(
-            getv(from_object, ['dynamic_retrieval_config']), to_object
-        ),
-    )
+  if getv(from_object, ['auth_config']) is not None:
+    raise ValueError('auth_config parameter is not supported in Gemini API.')
+
+  if getv(from_object, ['enable_widget']) is not None:
+    setv(to_object, ['enableWidget'], getv(from_object, ['enable_widget']))
 
   return to_object
 
@@ -1223,40 +1035,19 @@ def _GoogleSearch_to_mldev(
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
-  if getv(from_object, ['time_range_filter']) is not None:
-    setv(
-        to_object,
-        ['timeRangeFilter'],
-        _Interval_to_mldev(getv(from_object, ['time_range_filter']), to_object),
-    )
-
   if getv(from_object, ['exclude_domains']) is not None:
     raise ValueError(
         'exclude_domains parameter is not supported in Gemini API.'
     )
 
-  return to_object
-
-
-def _InlinedEmbedContentResponse_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['response']) is not None:
-    setv(
-        to_object,
-        ['response'],
-        _SingleEmbedContentResponse_from_mldev(
-            getv(from_object, ['response']), to_object
-        ),
+  if getv(from_object, ['blocking_confidence']) is not None:
+    raise ValueError(
+        'blocking_confidence parameter is not supported in Gemini API.'
     )
 
-  if getv(from_object, ['error']) is not None:
+  if getv(from_object, ['time_range_filter']) is not None:
     setv(
-        to_object,
-        ['error'],
-        _JobError_from_mldev(getv(from_object, ['error']), to_object),
+        to_object, ['timeRangeFilter'], getv(from_object, ['time_range_filter'])
     )
 
   return to_object
@@ -1285,12 +1076,17 @@ def _InlinedRequest_to_mldev(
         ],
     )
 
+  if getv(from_object, ['metadata']) is not None:
+    setv(to_object, ['metadata'], getv(from_object, ['metadata']))
+
   if getv(from_object, ['config']) is not None:
     setv(
         to_object,
         ['request', 'generationConfig'],
         _GenerateContentConfig_to_mldev(
-            api_client, getv(from_object, ['config']), to_object
+            api_client,
+            getv(from_object, ['config']),
+            getv(to_object, ['request'], default_value={}),
         ),
     )
 
@@ -1312,73 +1108,7 @@ def _InlinedResponse_from_mldev(
     )
 
   if getv(from_object, ['error']) is not None:
-    setv(
-        to_object,
-        ['error'],
-        _JobError_from_mldev(getv(from_object, ['error']), to_object),
-    )
-
-  return to_object
-
-
-def _Interval_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['start_time']) is not None:
-    setv(to_object, ['startTime'], getv(from_object, ['start_time']))
-
-  if getv(from_object, ['end_time']) is not None:
-    setv(to_object, ['endTime'], getv(from_object, ['end_time']))
-
-  return to_object
-
-
-def _JobError_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['details']) is not None:
-    setv(to_object, ['details'], getv(from_object, ['details']))
-
-  if getv(from_object, ['code']) is not None:
-    setv(to_object, ['code'], getv(from_object, ['code']))
-
-  if getv(from_object, ['message']) is not None:
-    setv(to_object, ['message'], getv(from_object, ['message']))
-
-  return to_object
-
-
-def _JobError_from_vertex(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['details']) is not None:
-    setv(to_object, ['details'], getv(from_object, ['details']))
-
-  if getv(from_object, ['code']) is not None:
-    setv(to_object, ['code'], getv(from_object, ['code']))
-
-  if getv(from_object, ['message']) is not None:
-    setv(to_object, ['message'], getv(from_object, ['message']))
-
-  return to_object
-
-
-def _LatLng_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['latitude']) is not None:
-    setv(to_object, ['latitude'], getv(from_object, ['latitude']))
-
-  if getv(from_object, ['longitude']) is not None:
-    setv(to_object, ['longitude'], getv(from_object, ['longitude']))
+    setv(to_object, ['error'], getv(from_object, ['error']))
 
   return to_object
 
@@ -1505,138 +1235,13 @@ def _ListBatchJobsResponse_from_vertex(
   return to_object
 
 
-def _MultiSpeakerVoiceConfig_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['speaker_voice_configs']) is not None:
-    setv(
-        to_object,
-        ['speakerVoiceConfigs'],
-        [
-            _SpeakerVoiceConfig_to_mldev(item, to_object)
-            for item in getv(from_object, ['speaker_voice_configs'])
-        ],
-    )
-
-  return to_object
-
-
-def _Part_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['videoMetadata']) is not None:
-    setv(
-        to_object,
-        ['video_metadata'],
-        _VideoMetadata_from_mldev(
-            getv(from_object, ['videoMetadata']), to_object
-        ),
-    )
-
-  if getv(from_object, ['thought']) is not None:
-    setv(to_object, ['thought'], getv(from_object, ['thought']))
-
-  if getv(from_object, ['inlineData']) is not None:
-    setv(
-        to_object,
-        ['inline_data'],
-        _Blob_from_mldev(getv(from_object, ['inlineData']), to_object),
-    )
-
-  if getv(from_object, ['fileData']) is not None:
-    setv(
-        to_object,
-        ['file_data'],
-        _FileData_from_mldev(getv(from_object, ['fileData']), to_object),
-    )
-
-  if getv(from_object, ['thoughtSignature']) is not None:
-    setv(
-        to_object,
-        ['thought_signature'],
-        getv(from_object, ['thoughtSignature']),
-    )
-
-  if getv(from_object, ['functionCall']) is not None:
-    setv(
-        to_object,
-        ['function_call'],
-        _FunctionCall_from_mldev(
-            getv(from_object, ['functionCall']), to_object
-        ),
-    )
-
-  if getv(from_object, ['codeExecutionResult']) is not None:
-    setv(
-        to_object,
-        ['code_execution_result'],
-        getv(from_object, ['codeExecutionResult']),
-    )
-
-  if getv(from_object, ['executableCode']) is not None:
-    setv(to_object, ['executable_code'], getv(from_object, ['executableCode']))
-
-  if getv(from_object, ['functionResponse']) is not None:
-    setv(
-        to_object,
-        ['function_response'],
-        getv(from_object, ['functionResponse']),
-    )
-
-  if getv(from_object, ['text']) is not None:
-    setv(to_object, ['text'], getv(from_object, ['text']))
-
-  return to_object
-
-
 def _Part_to_mldev(
     from_object: Union[dict[str, Any], object],
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
-  if getv(from_object, ['video_metadata']) is not None:
-    setv(
-        to_object,
-        ['videoMetadata'],
-        _VideoMetadata_to_mldev(
-            getv(from_object, ['video_metadata']), to_object
-        ),
-    )
-
-  if getv(from_object, ['thought']) is not None:
-    setv(to_object, ['thought'], getv(from_object, ['thought']))
-
-  if getv(from_object, ['inline_data']) is not None:
-    setv(
-        to_object,
-        ['inlineData'],
-        _Blob_to_mldev(getv(from_object, ['inline_data']), to_object),
-    )
-
-  if getv(from_object, ['file_data']) is not None:
-    setv(
-        to_object,
-        ['fileData'],
-        _FileData_to_mldev(getv(from_object, ['file_data']), to_object),
-    )
-
-  if getv(from_object, ['thought_signature']) is not None:
-    setv(
-        to_object,
-        ['thoughtSignature'],
-        getv(from_object, ['thought_signature']),
-    )
-
   if getv(from_object, ['function_call']) is not None:
-    setv(
-        to_object,
-        ['functionCall'],
-        _FunctionCall_to_mldev(getv(from_object, ['function_call']), to_object),
-    )
+    setv(to_object, ['functionCall'], getv(from_object, ['function_call']))
 
   if getv(from_object, ['code_execution_result']) is not None:
     setv(
@@ -1648,6 +1253,13 @@ def _Part_to_mldev(
   if getv(from_object, ['executable_code']) is not None:
     setv(to_object, ['executableCode'], getv(from_object, ['executable_code']))
 
+  if getv(from_object, ['file_data']) is not None:
+    setv(
+        to_object,
+        ['fileData'],
+        _FileData_to_mldev(getv(from_object, ['file_data']), to_object),
+    )
+
   if getv(from_object, ['function_response']) is not None:
     setv(
         to_object,
@@ -1655,37 +1267,28 @@ def _Part_to_mldev(
         getv(from_object, ['function_response']),
     )
 
+  if getv(from_object, ['inline_data']) is not None:
+    setv(
+        to_object,
+        ['inlineData'],
+        _Blob_to_mldev(getv(from_object, ['inline_data']), to_object),
+    )
+
   if getv(from_object, ['text']) is not None:
     setv(to_object, ['text'], getv(from_object, ['text']))
 
-  return to_object
+  if getv(from_object, ['thought']) is not None:
+    setv(to_object, ['thought'], getv(from_object, ['thought']))
 
-
-def _PrebuiltVoiceConfig_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['voice_name']) is not None:
-    setv(to_object, ['voiceName'], getv(from_object, ['voice_name']))
-
-  return to_object
-
-
-def _RetrievalConfig_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['lat_lng']) is not None:
+  if getv(from_object, ['thought_signature']) is not None:
     setv(
         to_object,
-        ['latLng'],
-        _LatLng_to_mldev(getv(from_object, ['lat_lng']), to_object),
+        ['thoughtSignature'],
+        getv(from_object, ['thought_signature']),
     )
 
-  if getv(from_object, ['language_code']) is not None:
-    setv(to_object, ['languageCode'], getv(from_object, ['language_code']))
+  if getv(from_object, ['video_metadata']) is not None:
+    setv(to_object, ['videoMetadata'], getv(from_object, ['video_metadata']))
 
   return to_object
 
@@ -1695,228 +1298,14 @@ def _SafetySetting_to_mldev(
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
-  if getv(from_object, ['method']) is not None:
-    raise ValueError('method parameter is not supported in Gemini API.')
-
   if getv(from_object, ['category']) is not None:
     setv(to_object, ['category'], getv(from_object, ['category']))
 
+  if getv(from_object, ['method']) is not None:
+    raise ValueError('method parameter is not supported in Gemini API.')
+
   if getv(from_object, ['threshold']) is not None:
     setv(to_object, ['threshold'], getv(from_object, ['threshold']))
-
-  return to_object
-
-
-def _Schema_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['additional_properties']) is not None:
-    raise ValueError(
-        'additional_properties parameter is not supported in Gemini API.'
-    )
-
-  if getv(from_object, ['defs']) is not None:
-    raise ValueError('defs parameter is not supported in Gemini API.')
-
-  if getv(from_object, ['ref']) is not None:
-    raise ValueError('ref parameter is not supported in Gemini API.')
-
-  if getv(from_object, ['any_of']) is not None:
-    setv(to_object, ['anyOf'], getv(from_object, ['any_of']))
-
-  if getv(from_object, ['default']) is not None:
-    setv(to_object, ['default'], getv(from_object, ['default']))
-
-  if getv(from_object, ['description']) is not None:
-    setv(to_object, ['description'], getv(from_object, ['description']))
-
-  if getv(from_object, ['enum']) is not None:
-    setv(to_object, ['enum'], getv(from_object, ['enum']))
-
-  if getv(from_object, ['example']) is not None:
-    setv(to_object, ['example'], getv(from_object, ['example']))
-
-  if getv(from_object, ['format']) is not None:
-    setv(to_object, ['format'], getv(from_object, ['format']))
-
-  if getv(from_object, ['items']) is not None:
-    setv(to_object, ['items'], getv(from_object, ['items']))
-
-  if getv(from_object, ['max_items']) is not None:
-    setv(to_object, ['maxItems'], getv(from_object, ['max_items']))
-
-  if getv(from_object, ['max_length']) is not None:
-    setv(to_object, ['maxLength'], getv(from_object, ['max_length']))
-
-  if getv(from_object, ['max_properties']) is not None:
-    setv(to_object, ['maxProperties'], getv(from_object, ['max_properties']))
-
-  if getv(from_object, ['maximum']) is not None:
-    setv(to_object, ['maximum'], getv(from_object, ['maximum']))
-
-  if getv(from_object, ['min_items']) is not None:
-    setv(to_object, ['minItems'], getv(from_object, ['min_items']))
-
-  if getv(from_object, ['min_length']) is not None:
-    setv(to_object, ['minLength'], getv(from_object, ['min_length']))
-
-  if getv(from_object, ['min_properties']) is not None:
-    setv(to_object, ['minProperties'], getv(from_object, ['min_properties']))
-
-  if getv(from_object, ['minimum']) is not None:
-    setv(to_object, ['minimum'], getv(from_object, ['minimum']))
-
-  if getv(from_object, ['nullable']) is not None:
-    setv(to_object, ['nullable'], getv(from_object, ['nullable']))
-
-  if getv(from_object, ['pattern']) is not None:
-    setv(to_object, ['pattern'], getv(from_object, ['pattern']))
-
-  if getv(from_object, ['properties']) is not None:
-    setv(to_object, ['properties'], getv(from_object, ['properties']))
-
-  if getv(from_object, ['property_ordering']) is not None:
-    setv(
-        to_object,
-        ['propertyOrdering'],
-        getv(from_object, ['property_ordering']),
-    )
-
-  if getv(from_object, ['required']) is not None:
-    setv(to_object, ['required'], getv(from_object, ['required']))
-
-  if getv(from_object, ['title']) is not None:
-    setv(to_object, ['title'], getv(from_object, ['title']))
-
-  if getv(from_object, ['type']) is not None:
-    setv(to_object, ['type'], getv(from_object, ['type']))
-
-  return to_object
-
-
-def _SingleEmbedContentResponse_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['embedding']) is not None:
-    setv(
-        to_object,
-        ['embedding'],
-        _ContentEmbedding_from_mldev(
-            getv(from_object, ['embedding']), to_object
-        ),
-    )
-
-  if getv(from_object, ['tokenCount']) is not None:
-    setv(to_object, ['token_count'], getv(from_object, ['tokenCount']))
-
-  return to_object
-
-
-def _SpeakerVoiceConfig_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['speaker']) is not None:
-    setv(to_object, ['speaker'], getv(from_object, ['speaker']))
-
-  if getv(from_object, ['voice_config']) is not None:
-    setv(
-        to_object,
-        ['voiceConfig'],
-        _VoiceConfig_to_mldev(getv(from_object, ['voice_config']), to_object),
-    )
-
-  return to_object
-
-
-def _SpeechConfig_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['voice_config']) is not None:
-    setv(
-        to_object,
-        ['voiceConfig'],
-        _VoiceConfig_to_mldev(getv(from_object, ['voice_config']), to_object),
-    )
-
-  if getv(from_object, ['multi_speaker_voice_config']) is not None:
-    setv(
-        to_object,
-        ['multiSpeakerVoiceConfig'],
-        _MultiSpeakerVoiceConfig_to_mldev(
-            getv(from_object, ['multi_speaker_voice_config']), to_object
-        ),
-    )
-
-  if getv(from_object, ['language_code']) is not None:
-    setv(to_object, ['languageCode'], getv(from_object, ['language_code']))
-
-  return to_object
-
-
-def _ThinkingConfig_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['include_thoughts']) is not None:
-    setv(
-        to_object, ['includeThoughts'], getv(from_object, ['include_thoughts'])
-    )
-
-  if getv(from_object, ['thinking_budget']) is not None:
-    setv(to_object, ['thinkingBudget'], getv(from_object, ['thinking_budget']))
-
-  return to_object
-
-
-def _ToolComputerUse_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['environment']) is not None:
-    setv(to_object, ['environment'], getv(from_object, ['environment']))
-
-  if getv(from_object, ['excluded_predefined_functions']) is not None:
-    setv(
-        to_object,
-        ['excludedPredefinedFunctions'],
-        getv(from_object, ['excluded_predefined_functions']),
-    )
-
-  return to_object
-
-
-def _ToolConfig_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['function_calling_config']) is not None:
-    setv(
-        to_object,
-        ['functionCallingConfig'],
-        _FunctionCallingConfig_to_mldev(
-            getv(from_object, ['function_calling_config']), to_object
-        ),
-    )
-
-  if getv(from_object, ['retrieval_config']) is not None:
-    setv(
-        to_object,
-        ['retrievalConfig'],
-        _RetrievalConfig_to_mldev(
-            getv(from_object, ['retrieval_config']), to_object
-        ),
-    )
 
   return to_object
 
@@ -1930,14 +1319,36 @@ def _Tool_to_mldev(
     setv(
         to_object,
         ['functionDeclarations'],
-        [
-            _FunctionDeclaration_to_mldev(item, to_object)
-            for item in getv(from_object, ['function_declarations'])
-        ],
+        [item for item in getv(from_object, ['function_declarations'])],
     )
 
   if getv(from_object, ['retrieval']) is not None:
     raise ValueError('retrieval parameter is not supported in Gemini API.')
+
+  if getv(from_object, ['google_search_retrieval']) is not None:
+    setv(
+        to_object,
+        ['googleSearchRetrieval'],
+        getv(from_object, ['google_search_retrieval']),
+    )
+
+  if getv(from_object, ['google_maps']) is not None:
+    setv(
+        to_object,
+        ['googleMaps'],
+        _GoogleMaps_to_mldev(getv(from_object, ['google_maps']), to_object),
+    )
+
+  if getv(from_object, ['computer_use']) is not None:
+    setv(to_object, ['computerUse'], getv(from_object, ['computer_use']))
+
+  if getv(from_object, ['code_execution']) is not None:
+    setv(to_object, ['codeExecution'], getv(from_object, ['code_execution']))
+
+  if getv(from_object, ['enterprise_web_search']) is not None:
+    raise ValueError(
+        'enterprise_web_search parameter is not supported in Gemini API.'
+    )
 
   if getv(from_object, ['google_search']) is not None:
     setv(
@@ -1946,137 +1357,8 @@ def _Tool_to_mldev(
         _GoogleSearch_to_mldev(getv(from_object, ['google_search']), to_object),
     )
 
-  if getv(from_object, ['google_search_retrieval']) is not None:
-    setv(
-        to_object,
-        ['googleSearchRetrieval'],
-        _GoogleSearchRetrieval_to_mldev(
-            getv(from_object, ['google_search_retrieval']), to_object
-        ),
-    )
-
-  if getv(from_object, ['enterprise_web_search']) is not None:
-    raise ValueError(
-        'enterprise_web_search parameter is not supported in Gemini API.'
-    )
-
-  if getv(from_object, ['google_maps']) is not None:
-    raise ValueError('google_maps parameter is not supported in Gemini API.')
-
   if getv(from_object, ['url_context']) is not None:
-    setv(
-        to_object,
-        ['urlContext'],
-        _UrlContext_to_mldev(getv(from_object, ['url_context']), to_object),
-    )
-
-  if getv(from_object, ['computer_use']) is not None:
-    setv(
-        to_object,
-        ['computerUse'],
-        _ToolComputerUse_to_mldev(
-            getv(from_object, ['computer_use']), to_object
-        ),
-    )
-
-  if getv(from_object, ['code_execution']) is not None:
-    setv(to_object, ['codeExecution'], getv(from_object, ['code_execution']))
-
-  return to_object
-
-
-def _UrlContextMetadata_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['urlMetadata']) is not None:
-    setv(
-        to_object,
-        ['url_metadata'],
-        [
-            _UrlMetadata_from_mldev(item, to_object)
-            for item in getv(from_object, ['urlMetadata'])
-        ],
-    )
-
-  return to_object
-
-
-def _UrlContext_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-
-  return to_object
-
-
-def _UrlMetadata_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['retrievedUrl']) is not None:
-    setv(to_object, ['retrieved_url'], getv(from_object, ['retrievedUrl']))
-
-  if getv(from_object, ['urlRetrievalStatus']) is not None:
-    setv(
-        to_object,
-        ['url_retrieval_status'],
-        getv(from_object, ['urlRetrievalStatus']),
-    )
-
-  return to_object
-
-
-def _VideoMetadata_from_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['fps']) is not None:
-    setv(to_object, ['fps'], getv(from_object, ['fps']))
-
-  if getv(from_object, ['endOffset']) is not None:
-    setv(to_object, ['end_offset'], getv(from_object, ['endOffset']))
-
-  if getv(from_object, ['startOffset']) is not None:
-    setv(to_object, ['start_offset'], getv(from_object, ['startOffset']))
-
-  return to_object
-
-
-def _VideoMetadata_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['fps']) is not None:
-    setv(to_object, ['fps'], getv(from_object, ['fps']))
-
-  if getv(from_object, ['end_offset']) is not None:
-    setv(to_object, ['endOffset'], getv(from_object, ['end_offset']))
-
-  if getv(from_object, ['start_offset']) is not None:
-    setv(to_object, ['startOffset'], getv(from_object, ['start_offset']))
-
-  return to_object
-
-
-def _VoiceConfig_to_mldev(
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['prebuilt_voice_config']) is not None:
-    setv(
-        to_object,
-        ['prebuiltVoiceConfig'],
-        _PrebuiltVoiceConfig_to_mldev(
-            getv(from_object, ['prebuilt_voice_config']), to_object
-        ),
-    )
+    setv(to_object, ['urlContext'], getv(from_object, ['url_context']))
 
   return to_object
 
@@ -2136,12 +1418,12 @@ class Batches(_api_module.BaseModule):
         'post', path, request_dict, http_options
     )
 
-    response_dict = '' if not response.body else json.loads(response.body)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if self._api_client.vertexai:
       response_dict = _BatchJob_from_vertex(response_dict)
 
-    else:
+    if not self._api_client.vertexai:
       response_dict = _BatchJob_from_mldev(response_dict)
 
     return_value = types.BatchJob._from_response(
@@ -2199,7 +1481,7 @@ class Batches(_api_module.BaseModule):
         'post', path, request_dict, http_options
     )
 
-    response_dict = '' if not response.body else json.loads(response.body)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if not self._api_client.vertexai:
       response_dict = _BatchJob_from_mldev(response_dict)
@@ -2276,12 +1558,12 @@ class Batches(_api_module.BaseModule):
 
     response = self._api_client.request('get', path, request_dict, http_options)
 
-    response_dict = '' if not response.body else json.loads(response.body)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if self._api_client.vertexai:
       response_dict = _BatchJob_from_vertex(response_dict)
 
-    else:
+    if not self._api_client.vertexai:
       response_dict = _BatchJob_from_mldev(response_dict)
 
     return_value = types.BatchJob._from_response(
@@ -2400,12 +1682,12 @@ class Batches(_api_module.BaseModule):
 
     response = self._api_client.request('get', path, request_dict, http_options)
 
-    response_dict = '' if not response.body else json.loads(response.body)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if self._api_client.vertexai:
       response_dict = _ListBatchJobsResponse_from_vertex(response_dict)
 
-    else:
+    if not self._api_client.vertexai:
       response_dict = _ListBatchJobsResponse_from_mldev(response_dict)
 
     return_value = types.ListBatchJobsResponse._from_response(
@@ -2485,12 +1767,12 @@ class Batches(_api_module.BaseModule):
         'delete', path, request_dict, http_options
     )
 
-    response_dict = '' if not response.body else json.loads(response.body)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if self._api_client.vertexai:
       response_dict = _DeleteResourceJob_from_vertex(response_dict)
 
-    else:
+    if not self._api_client.vertexai:
       response_dict = _DeleteResourceJob_from_mldev(response_dict)
 
     return_value = types.DeleteResourceJob._from_response(
@@ -2541,36 +1823,11 @@ class Batches(_api_module.BaseModule):
         config=config,
     )
 
-    http_options: Optional[types.HttpOptions] = None
-    if (
-        parameter_model.config is not None
-        and parameter_model.config.http_options is not None
-    ):
-      http_options = parameter_model.config.http_options
-
     if self._api_client.vertexai:
       config = _extra_utils.format_destination(src, parameter_model.config)
       return self._create(model=model, src=src, config=config)
-    elif src.inlined_requests is None:
+    else:
       return self._create(model=model, src=src, config=config)
-
-    path, request_dict = _create_inlined_generate_content_request_dict(
-        self._api_client, parameter_model
-    )
-
-    response = self._api_client.request(
-        'post', path, request_dict, http_options
-    )
-
-    response_dict = '' if not response.body else json.loads(response.body)
-    response_dict = _BatchJob_from_mldev(response_dict)
-
-    return_value = types.BatchJob._from_response(
-        response=response_dict, kwargs=parameter_model.model_dump()
-    )
-
-    self._api_client._verify_response(return_value)
-    return return_value
 
   def create_embeddings(
       self,
@@ -2617,35 +1874,10 @@ class Batches(_api_module.BaseModule):
         config=config,
     )
 
-    http_options: Optional[types.HttpOptions] = None
-    if (
-        parameter_model.config is not None
-        and parameter_model.config.http_options is not None
-    ):
-      http_options = parameter_model.config.http_options
-
     if self._api_client.vertexai:
       raise ValueError('Vertex AI does not support batches.create_embeddings.')
-    elif src.inlined_requests is None:
+    else:
       return self._create_embeddings(model=model, src=src, config=config)
-
-    path, request_dict = _create_inlined_embedding_request_dict(
-        self._api_client, parameter_model
-    )
-
-    response = self._api_client.request(
-        'post', path, request_dict, http_options
-    )
-
-    response_dict = '' if not response.body else json.loads(response.body)
-    response_dict = _BatchJob_from_mldev(response_dict)
-
-    return_value = types.BatchJob._from_response(
-        response=response_dict, kwargs=parameter_model.model_dump()
-    )
-
-    self._api_client._verify_response(return_value)
-    return return_value
 
   def list(
       self, *, config: Optional[types.ListBatchJobsConfigOrDict] = None
@@ -2732,12 +1964,12 @@ class AsyncBatches(_api_module.BaseModule):
         'post', path, request_dict, http_options
     )
 
-    response_dict = '' if not response.body else json.loads(response.body)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if self._api_client.vertexai:
       response_dict = _BatchJob_from_vertex(response_dict)
 
-    else:
+    if not self._api_client.vertexai:
       response_dict = _BatchJob_from_mldev(response_dict)
 
     return_value = types.BatchJob._from_response(
@@ -2795,7 +2027,7 @@ class AsyncBatches(_api_module.BaseModule):
         'post', path, request_dict, http_options
     )
 
-    response_dict = '' if not response.body else json.loads(response.body)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if not self._api_client.vertexai:
       response_dict = _BatchJob_from_mldev(response_dict)
@@ -2874,12 +2106,12 @@ class AsyncBatches(_api_module.BaseModule):
         'get', path, request_dict, http_options
     )
 
-    response_dict = '' if not response.body else json.loads(response.body)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if self._api_client.vertexai:
       response_dict = _BatchJob_from_vertex(response_dict)
 
-    else:
+    if not self._api_client.vertexai:
       response_dict = _BatchJob_from_mldev(response_dict)
 
     return_value = types.BatchJob._from_response(
@@ -3000,12 +2232,12 @@ class AsyncBatches(_api_module.BaseModule):
         'get', path, request_dict, http_options
     )
 
-    response_dict = '' if not response.body else json.loads(response.body)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if self._api_client.vertexai:
       response_dict = _ListBatchJobsResponse_from_vertex(response_dict)
 
-    else:
+    if not self._api_client.vertexai:
       response_dict = _ListBatchJobsResponse_from_mldev(response_dict)
 
     return_value = types.ListBatchJobsResponse._from_response(
@@ -3085,12 +2317,12 @@ class AsyncBatches(_api_module.BaseModule):
         'delete', path, request_dict, http_options
     )
 
-    response_dict = '' if not response.body else json.loads(response.body)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if self._api_client.vertexai:
       response_dict = _DeleteResourceJob_from_vertex(response_dict)
 
-    else:
+    if not self._api_client.vertexai:
       response_dict = _DeleteResourceJob_from_mldev(response_dict)
 
     return_value = types.DeleteResourceJob._from_response(
@@ -3140,36 +2372,11 @@ class AsyncBatches(_api_module.BaseModule):
         config=config,
     )
 
-    http_options: Optional[types.HttpOptions] = None
-    if (
-        parameter_model.config is not None
-        and parameter_model.config.http_options is not None
-    ):
-      http_options = parameter_model.config.http_options
-
     if self._api_client.vertexai:
       config = _extra_utils.format_destination(src, parameter_model.config)
       return await self._create(model=model, src=src, config=config)
-    elif src.inlined_requests is None:
+    else:
       return await self._create(model=model, src=src, config=config)
-
-    path, request_dict = _create_inlined_generate_content_request_dict(
-        self._api_client, parameter_model
-    )
-
-    response = await self._api_client.async_request(
-        'post', path, request_dict, http_options
-    )
-
-    response_dict = '' if not response.body else json.loads(response.body)
-    response_dict = _BatchJob_from_mldev(response_dict)
-
-    return_value = types.BatchJob._from_response(
-        response=response_dict, kwargs=parameter_model.model_dump()
-    )
-
-    self._api_client._verify_response(return_value)
-    return return_value
 
   async def create_embeddings(
       self,
@@ -3225,26 +2432,8 @@ class AsyncBatches(_api_module.BaseModule):
 
     if self._api_client.vertexai:
       raise ValueError('Vertex AI does not support batches.create_embeddings.')
-    elif src.inlined_requests is None:
+    else:
       return await self._create_embeddings(model=model, src=src, config=config)
-
-    path, request_dict = _create_inlined_embedding_request_dict(
-        self._api_client, parameter_model
-    )
-
-    response = await self._api_client.async_request(
-        'post', path, request_dict, http_options
-    )
-
-    response_dict = '' if not response.body else json.loads(response.body)
-    response_dict = _BatchJob_from_mldev(response_dict)
-
-    return_value = types.BatchJob._from_response(
-        response=response_dict, kwargs=parameter_model.model_dump()
-    )
-
-    self._api_client._verify_response(return_value)
-    return return_value
 
   async def list(
       self, *, config: Optional[types.ListBatchJobsConfigOrDict] = None
@@ -3275,100 +2464,3 @@ class AsyncBatches(_api_module.BaseModule):
         await self._list(config=config),
         config,
     )
-
-
-def _create_inlined_generate_content_request_dict(
-    client: BaseApiClient, parameter_model: types._CreateBatchJobParameters
-) -> tuple[str, dict[str, Any]]:
-  request_url_dict: Optional[dict[str, str]]
-
-  request_dict: dict[str, Any] = _CreateBatchJobParameters_to_mldev(
-      client, parameter_model
-  )
-
-  request_url_dict = request_dict.get('_url')
-  if request_url_dict:
-    path = '{model}:batchGenerateContent'.format_map(request_url_dict)
-  else:
-    path = '{model}:batchGenerateContent'
-  query_params = request_dict.get('_query')
-  if query_params:
-    path = f'{path}?{urlencode(query_params)}'
-  request_dict.pop('config', None)
-
-  request_dict = _common.convert_to_dict(request_dict)
-  request_dict = _common.encode_unserializable_types(request_dict)
-  # Move system instruction to 'request':
-  # {'systemInstruction': system_instruction}
-  requests = []
-  batch_dict = request_dict.get('batch')
-  if batch_dict and isinstance(batch_dict, dict):
-    input_config_dict = batch_dict.get('inputConfig')
-    if input_config_dict and isinstance(input_config_dict, dict):
-      requests_dict = input_config_dict.get('requests')
-      if requests_dict and isinstance(requests_dict, dict):
-        requests = requests_dict.get('requests')
-  new_requests = []
-  if requests:
-    for req in requests:
-      if req.get('systemInstruction'):
-        value = req.pop('systemInstruction')
-        req['request'].update({'systemInstruction': value})
-      new_requests.append(req)
-  request_dict['batch']['inputConfig']['requests'][  # type: ignore
-      'requests'
-  ] = new_requests
-  return path, request_dict
-
-
-def _create_inlined_embedding_request_dict(
-    client: BaseApiClient,
-    parameter_model: types._CreateEmbeddingsBatchJobParameters,
-) -> tuple[str, dict[str, Any]]:
-  src = parameter_model.src
-  if not isinstance(src, types.EmbeddingsBatchJobSource):
-    raise ValueError(f'Invalid batch job source: {src}.')
-
-  request_url_dict: Optional[dict[str, str]]
-
-  request_dict: dict[str, Any] = _CreateEmbeddingsBatchJobParameters_to_mldev(
-      client, parameter_model
-  )
-
-  request_url_dict = request_dict.get('_url')
-  if request_url_dict:
-    path = '{model}:asyncBatchEmbedContent'.format_map(request_url_dict)
-  else:
-    path = '{model}:asyncBatchEmbedContent'
-  query_params = request_dict.get('_query')
-  if query_params:
-    path = f'{path}?{urlencode(query_params)}'
-
-  request_dict.pop('config', None)
-  request_dict.get('batch', {}).get('inputConfig', {}).get('requests', {}).pop(
-      'config', None
-  )
-
-  request_dict = _common.convert_to_dict(request_dict)
-  request_dict = _common.encode_unserializable_types(request_dict)
-
-  requests = []
-  batch_dict = request_dict.get('batch')
-  if batch_dict and isinstance(batch_dict, dict):
-    input_config_dict = batch_dict.get('inputConfig')
-    if input_config_dict and isinstance(input_config_dict, dict):
-      requests_dict = input_config_dict.get('requests')
-      if requests_dict and isinstance(requests_dict, dict):
-        requests = requests_dict.get('requests')
-  new_requests = []
-  if requests:
-    for req in requests:
-      for k in list(req.keys()):
-        if k != 'request':
-          req['request'][k] = req.pop(k)
-      new_requests.append(req)
-  request_dict['batch']['inputConfig']['requests'][  # type: ignore
-      'requests'
-  ] = new_requests
-
-  return path, request_dict
