@@ -397,4 +397,7 @@ class Client:
     self.close()
 
   def __del__(self) -> None:
-    self.close()
+    try:
+      self.close()
+    except Exception:
+      pass
