@@ -20,21 +20,35 @@ from ... import pytest_helper
 
 
 def create_update_get(client, parameters):
-  cache = client.caches.create(
-      model=parameters.model,
-      config=parameters.config,
-  )
+  if client.vertexai:
+    cache = client.caches.create(
+        model=parameters.model,
+        config=parameters.config,
+    )
+  else:
+    # For MLDev, we must create the file before creating the cache.
+    file = client.files.upload(file="tests/data/a-man-and-a-dog.png")
+    config = genai_types.CreateCachedContentConfig(
+        contents=[
+            genai_types.Part.from_uri(
+                file_uri=file.uri,
+                mime_type=file.mime_type,
+            )
+        ]
+        * 5
+    )
+    cache = client.caches.create(model=parameters.model, config=config)
+
   updated_cache = client.caches.update(
       name=cache.name,
-      config=genai_types.UpdateCachedContentConfig(ttl="3600s"),
+      config=genai_types.UpdateCachedContentConfig(ttl="7200s"),
   )
-  got_cache = client.caches.get(name=updated_cache.name)
-  client.caches.delete(name=got_cache.name)
+  _ = client.caches.get(name=updated_cache.name)
 
 
 test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
-        name="test_create_update_get_vertex",
+        name="test_create_update_get",
         parameters=genai_types._CreateCachedContentParameters(
             model="gemini-2.5-flash",
             config=genai_types.CreateCachedContentConfig(
@@ -44,7 +58,6 @@ test_table: list[pytest_helper.TestTableItem] = [
                 )] * 5
             ),
         ),
-        exception_if_mldev="unsupported",
     ),
 ]
 
