@@ -35,7 +35,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name="test_dataset_gcs_uri",
         parameters=genai_types.CreateTuningJobParameters(
-            base_model="gemini-1.5-pro-002",
+            base_model="gemini-2.5-flash",
             training_dataset=genai_types.TuningDataset(
                 gcs_uri="gs://cloud-samples-data/ai-platform/generative_ai/gemini-1_5/text/sft_train_data.jsonl",
             ),
@@ -66,6 +66,37 @@ test_table: list[pytest_helper.TestTableItem] = [
         exception_if_mldev="is not supported in Gemini API",
     ),
     pytest_helper.TestTableItem(
+        name="test_tune_simple_dpo",
+        parameters=genai_types.CreateTuningJobParameters(
+            base_model="gemini-2.5-flash",
+            training_dataset=genai_types.TuningDataset(
+                gcs_uri="gs://cloud-samples-data/ai-platform/generative_ai/gemini-1_5/text/sft_train_data.jsonl",
+            ),
+            config=genai_types.CreateTuningJobConfig(
+                tuned_model_display_name="Model display name",
+                epoch_count=1,
+                method="PREFERENCE_TUNING",
+            ),
+        ),
+        exception_if_mldev="parameter is not supported in Gemini API.",
+    ),
+    pytest_helper.TestTableItem(
+        name="test_tune_dpo_with_beta",
+        parameters=genai_types.CreateTuningJobParameters(
+            base_model="gemini-2.5-flash",
+            training_dataset=genai_types.TuningDataset(
+                gcs_uri="gs://cloud-samples-data/ai-platform/generative_ai/gemini-1_5/text/sft_train_data.jsonl",
+            ),
+            config=genai_types.CreateTuningJobConfig(
+                tuned_model_display_name="Model display name",
+                epoch_count=1,
+                method=genai_types.TuningMethod.PREFERENCE_TUNING,
+                beta=0.5,
+            ),
+        ),
+        exception_if_mldev="parameter is not supported in Gemini API.",
+    ),
+    pytest_helper.TestTableItem(
         name="test_non_pretuned_model_with_checkpoint_id",
         parameters=genai_types.CreateTuningJobParameters(
             base_model="gemini-2.5-flash",
@@ -81,7 +112,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name="test_dataset_gcs_uri_all_parameters",
         parameters=genai_types.CreateTuningJobParameters(
-            base_model="gemini-1.5-pro-002",
+            base_model="gemini-2.5-flash",
             training_dataset=genai_types.TuningDataset(
                 gcs_uri="gs://cloud-samples-data/ai-platform/generative_ai/gemini-1_5/text/sft_train_data.jsonl",
             ),
@@ -90,69 +121,19 @@ test_table: list[pytest_helper.TestTableItem] = [
                 epoch_count=1,
                 learning_rate_multiplier=1.0,
                 adapter_size="ADAPTER_SIZE_ONE",
-                validation_dataset=genai_types.TuningDataset(
+                validation_dataset=genai_types.TuningValidationDataset(
                     gcs_uri="gs://cloud-samples-data/ai-platform/generative_ai/gemini-1_5/text/sft_validation_data.jsonl",
                 ),
-                # Not supported in Vertex AI
-                # batch_size=4,
-                # learning_rate=0.01,
             ),
         ),
         exception_if_mldev="gcs_uri parameter is not supported in Gemini API.",
-    ),
-    pytest_helper.TestTableItem(
-        name="test_dataset_gcs_uri_parameters_unsupported_by_vertex",
-        parameters=genai_types.CreateTuningJobParameters(
-            base_model="gemini-1.5-pro-002",
-            training_dataset=genai_types.TuningDataset(
-                gcs_uri="gs://cloud-samples-data/ai-platform/generative_ai/gemini-1_5/text/sft_train_data.jsonl",
-            ),
-            config=genai_types.CreateTuningJobConfig(
-                # Not supported in Vertex AI
-                batch_size=4,
-                learning_rate=0.01,
-            ),
-        ),
-        exception_if_vertex="batch_size parameter is not supported in Vertex AI.",
-        exception_if_mldev="gcs_uri parameter is not supported in Gemini API.",
-    ),
-    pytest_helper.TestTableItem(
-        name="test_dataset_examples_parameters_unsupported_by_mldev",
-        parameters=genai_types.CreateTuningJobParameters(
-            # Error: "models/gemini-1.5-pro-002 is not found for
-            # CREATE TUNED MODEL at API version v1beta."
-            # base_model="gemini-1.5-pro-002",
-            base_model="models/gemini-1.0-pro-001",
-            training_dataset=genai_types.TuningDataset(
-                examples=[
-                    genai_types.TuningExample(
-                        text_input=f"Input text {i}",
-                        output=f"Output text {i}",
-                    )
-                    for i in range(5)
-                ],
-            ),
-            # Required for MLDev:
-            # "Either tuned_model_id or display_name must be set."
-            config=genai_types.CreateTuningJobConfig(
-                tuned_model_display_name="Model display name",
-                # Not supported in MLDev
-                adapter_size="ADAPTER_SIZE_ONE",
-                # Generator issue: "validationDatasetUri": {}. See b/375079287
-                # validation_dataset=genai_types.TuningDataset(
-                #     gcs_uri="gs://cloud-samples-data/ai-platform/generative_ai/gemini-1_5/text/sft_validation_data.jsonl",
-                # ),
-            ),
-        ),
-        exception_if_mldev="adapter_size parameter is not supported in Gemini API.",
-        exception_if_vertex="examples parameter is not supported in Vertex AI.",
     ),
     pytest_helper.TestTableItem(
         name="test_dataset_vertex_dataset_resource",
         parameters=genai_types.CreateTuningJobParameters(
-            base_model="gemini-1.5-pro-002",
+            base_model="gemini-2.5-flash",
             training_dataset=genai_types.TuningDataset(
-                vertex_dataset_resource="projects/613165508263/locations/us-central1/datasets/8254568702121345024",
+                vertex_dataset_resource="projects/801452371447/locations/us-central1/datasets/5779918772206829568",
             ),
         ),
         exception_if_mldev="vertex_dataset_resource parameter is not supported in Gemini API.",
@@ -160,22 +141,19 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name="test_dataset_dataset_resource_all_parameters",
         parameters=genai_types.CreateTuningJobParameters(
-            base_model="gemini-1.5-pro-002",
+            base_model="gemini-2.5-flash",
             training_dataset=genai_types.TuningDataset(
-                vertex_dataset_resource="projects/613165508263/locations/us-central1/datasets/8254568702121345024",
+                vertex_dataset_resource="projects/801452371447/locations/us-central1/datasets/5779918772206829568",
             ),
             config=genai_types.CreateTuningJobConfig(
                 tuned_model_display_name="Model display name",
                 epoch_count=1,
                 learning_rate_multiplier=1.0,
                 adapter_size="ADAPTER_SIZE_ONE",
-                validation_dataset=genai_types.TuningDataset(
-                    vertex_dataset_resource="projects/613165508263/locations/us-central1/datasets/5556912525326417920",
+                validation_dataset=genai_types.TuningValidationDataset(
+                    vertex_dataset_resource="projects/801452371447/locations/us-central1/datasets/1168232753779441664",
                 ),
                 labels={"testlabelkey": "testlabelvalue"},
-                # Not supported in Vertex AI
-                # batch_size=4,
-                # learning_rate=0.01,
             ),
         ),
         exception_if_mldev="vertex_dataset_resource parameter is not supported in Gemini API.",
@@ -208,7 +186,7 @@ def test_eval_config(client):
           epoch_count=1,
           learning_rate_multiplier=1.0,
           adapter_size="ADAPTER_SIZE_ONE",
-          validation_dataset=genai_types.TuningDataset(
+          validation_dataset=genai_types.TuningValidationDataset(
               gcs_uri="gs://cloud-samples-data/ai-platform/generative_ai/gemini-2_0/text/sft_validation_data.jsonl"
           ),
           evaluation_config=evaluation_config,
@@ -255,7 +233,7 @@ def test_eval_config_with_metrics(client):
           epoch_count=1,
           learning_rate_multiplier=1.0,
           adapter_size="ADAPTER_SIZE_ONE",
-          validation_dataset=genai_types.TuningDataset(
+          validation_dataset=genai_types.TuningValidationDataset(
               gcs_uri="gs://cloud-samples-data/ai-platform/generative_ai/gemini-2_0/text/sft_validation_data.jsonl"
           ),
           evaluation_config=evaluation_config,
@@ -280,7 +258,7 @@ async def test_eval_config_async(client):
           epoch_count=1,
           learning_rate_multiplier=1.0,
           adapter_size="ADAPTER_SIZE_ONE",
-          validation_dataset=genai_types.TuningDataset(
+          validation_dataset=genai_types.TuningValidationDataset(
               gcs_uri="gs://cloud-samples-data/ai-platform/generative_ai/gemini-2_0/text/sft_validation_data.jsonl"
           ),
           evaluation_config=evaluation_config,
